moves support services into js-sdk
