+++ packages/linode-js-sdk/src/index.ts
@@ -2,6 +2,8 @@ export * from './account';
 
 export * from './profile';
 
+export * from './support';
+
 export * from './volumes';
 
 export * from './domains';
