+++ packages/linode-js-sdk/src/support/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+
+export * from './support';
+
+export * from './support.schema';
