--- packages/manager/src/services/support/index.ts
@@ -1,1 +1,0 @@
-export * from './support';
